Fetch the old-import git repo. Separate svncloner into a new file. Add doc.go.
