+++ cmd/gs-svn-import/doc.go
@@ -1,0 +1,3 @@
+// gs-svn-import is a tool for importing GNUstep projects from Subversion
+// into Git, preserving branches and preserving old import history.
+package main // import "github.com/gnustep/gitsvn-scripts/cmd/gs-svn-import"
